+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Please mark backwards incompatible changes with an exclamation mark at the start
 
 ## [Unreleased]
 
+### Changed
+- `GitilesHelper#gitiles_url` can now be called without a `path`. When no path
+  is given the method generates a link to the given `refspec` instead.
+
 ## [27.2.1] - 2025-03-14
 
 ### Fixed